decouple VariableTypeResolver and TypeToStringResolver
