+++ packages/NodeTypeResolver/src/NodeTypeResolver.php
@@ -7,17 +7,9 @@ use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Param;
 use PHPStan\Analyser\Scope;
-use PHPStan\Broker\Broker;
-use PHPStan\TrinaryLogic;
-use PHPStan\Type\IntersectionType;
-use PHPStan\Type\ObjectType;
-use PHPStan\Type\ThisType;
-use PHPStan\Type\Type;
-use PHPStan\Type\UnionType;
-use Rector\BetterPhpDocParser\NodeAnalyzer\DocBlockAnalyzer;
 use Rector\Node\Attribute;
 use Rector\NodeTypeResolver\Contract\PerNodeTypeResolver\PerNodeTypeResolverInterface;
-use Rector\NodeTypeResolver\Reflection\ClassReflectionTypesResolver;
+use Rector\NodeTypeResolver\PHPStan\Type\TypeToStringResolver;
 
 final class NodeTypeResolver
 {
@@ -27,28 +19,13 @@ final class NodeTypeResolver
     private $perNodeTypeResolvers = [];
 
     /**
-     * @var ClassReflectionTypesResolver
+     * @var TypeToStringResolver
      */
-    private $classReflectionTypesResolver;
+    private $typeToStringResolver;
 
-    /**
-     * @var DocBlockAnalyzer
-     */
-    private $docBlockAnalyzer;
-
-    /**
-     * @var Broker
-     */
-    private $broker;
-
-    public function __construct(
-        ClassReflectionTypesResolver $classReflectionTypesResolver,
-        DocBlockAnalyzer $docBlockAnalyzer,
-        Broker $broker
-    ) {
-        $this->classReflectionTypesResolver = $classReflectionTypesResolver;
-        $this->docBlockAnalyzer = $docBlockAnalyzer;
-        $this->broker = $broker;
+    public function __construct(TypeToStringResolver $typeToStringResolver)
+    {
+        $this->typeToStringResolver = $typeToStringResolver;
     }
 
     public function addPerNodeTypeResolver(PerNodeTypeResolverInterface $perNodeTypeResolver): void
@@ -74,11 +51,7 @@ final class NodeTypeResolver
             return [$node->type->toString()];
         }
 
-        if ($node instanceof Variable) {
-            return $this->resolveVariableNode($node, $nodeScope);
-        }
-
-        if ($node instanceof Expr) {
+        if ($node instanceof Expr && ! $node instanceof Variable) {
             return $this->resolveExprNode($node);
         }
 
@@ -100,81 +73,6 @@ final class NodeTypeResolver
 
         $type = $nodeScope->getType($exprNode);
 
-        return $this->resolveObjectTypesToStrings($type);
-    }
-
-    /**
-     * @todo make use of recursion
-     * @return string[]
-     */
-    private function resolveObjectTypesToStrings(Type $type): array
-    {
-        $types = [];
-
-        if ($type instanceof ObjectType) {
-            $types[] = $type->getClassName();
-        }
-
-        if ($type instanceof UnionType) {
-            foreach ($type->getTypes() as $type) {
-                if ($type instanceof ObjectType) {
-                    $types[] = $type->getClassName();
-                }
-            }
-        }
-
-        if ($type instanceof IntersectionType) {
-            foreach ($type->getTypes() as $type) {
-                if ($type instanceof ObjectType) {
-                    $types[] = $type->getClassName();
-                }
-            }
-        }
-
-        return $types;
-    }
-
-    /**
-     * @return string[]
-     */
-    private function resolveVariableNode(Variable $variableNode, Scope $nodeScope): array
-    {
-        $variableName = (string) $variableNode->name;
-
-        if ($nodeScope->hasVariableType($variableName) === TrinaryLogic::createYes()) {
-            $type = $nodeScope->getVariableType($variableName);
-
-            // this
-            if ($type instanceof ThisType) {
-                return $this->classReflectionTypesResolver->resolve($nodeScope->getClassReflection());
-            }
-
-            $types = $this->resolveObjectTypesToStrings($type);
-
-            // complete parents
-            foreach ($types as $type) {
-                $propertyClassReflection = $this->broker->getClass($type);
-                $types = array_merge($types, $this->classReflectionTypesResolver->resolve($propertyClassReflection));
-            }
-
-            return array_unique($types);
-        }
-
-        // get from annotation
-        $variableTypes = $this->docBlockAnalyzer->getVarTypes($variableNode);
-
-        foreach ($variableTypes as $i => $type) {
-            if (! class_exists($type)) {
-                unset($variableTypes[$i]);
-                continue;
-            }
-            $propertyClassReflection = $this->broker->getClass($type);
-            $variableTypes = array_merge(
-                $variableTypes,
-                $this->classReflectionTypesResolver->resolve($propertyClassReflection)
-            );
-        }
-
-        return array_unique($variableTypes);
+        return $this->typeToStringResolver->resolve($type);
     }
 }
